fix: retrieve API secrets in one API call (#397)
Update how the API's SSM ParameterStore params are retrieved
to use a single API call.  Also adds a check to cause the entry.sh
to fail if secrets are not retrieved.

This change is being made to try to avoid scenarios we have seen
where a lambda function gets initialized missing some of its
secrets.